requirements.txt
google-cloud-asset
google-cloud-container
google-cloud-iam
google-cloud-monitoring
google-cloud-storage
google-cloud-resource-manager
google-cloud-security-center
google-cloud-kms
google-cloud-audit-logs
google-cloud-compute
